More clean-up - Work on #37 - Units can now have multiple interactions - Fix some compiler warnings

+++ Model/Orders/Movement/ReconOrder.cs
@@ -36,7 +36,7 @@ namespace PanzerBlitz
 		{
 			Turns = Stream.ReadInt32();
 			ExitTile = (Tile)Objects[Stream.ReadInt32()];
-			_RemainingTurns = (int)Stream.ReadInt32();
+			_RemainingTurns = Stream.ReadInt32();
 		}
 
 		public void Serialize(SerializationOutputStream Stream)
